+++ Assets/UniGLTF/Runtime/UniGLTF/IO/ImporterContext.cs
@@ -105,50 +105,22 @@ namespace UniGLTF
 
         #endregion
 
-        IShaderStore m_shaderStore;
-        public IShaderStore ShaderStore
-        {
-            get
-            {
-                if (m_shaderStore == null)
-                {
-                    m_shaderStore = new ShaderStore(this);
-                }
-                return m_shaderStore;
-            }
-        }
-
-        MaterialImporter m_materialImporter;
-        protected void SetMaterialImporter(MaterialImporter importer)
-        {
-            m_materialImporter = importer;
-        }
-        public MaterialImporter MaterialImporter
-        {
-            get
-            {
-                if (m_materialImporter == null)
-                {
-                    m_materialImporter = new MaterialImporter(ShaderStore);
-                }
-                return m_materialImporter;
-            }
-        }
+        MaterialFactory m_materialFactory = new MaterialFactory();
+        public MaterialFactory MaterialFacotry => m_materialFactory;
 
         public ImporterContext(IShaderStore shaderStore)
         {
-            m_shaderStore = shaderStore;
+            m_materialFactory.ShaderStore = shaderStore;
         }
 
         public ImporterContext(MaterialImporter materialImporter)
         {
-            m_materialImporter = materialImporter;
+            m_materialFactory.MaterialImporter = materialImporter;
         }
 
         public ImporterContext()
         {
         }
-
         #region Source
 
         /// <summary>
@@ -451,50 +423,6 @@ namespace UniGLTF
             Root.name = Path.GetFileNameWithoutExtension(path);
         }
 
-        public virtual ITextureLoader CreateTextureLoader(int index)
-        {
-#if UNIGLTF_USE_WEBREQUEST_TEXTURELOADER
-            return new UnityWebRequestTextureLoader(index);
-#else
-            return new TextureLoader(index);
-#endif
-        }
-
-        public void CreateTextureItems(UnityPath imageBaseDir = default(UnityPath))
-        {
-            if (m_textures.Any())
-            {
-                return;
-            }
-
-            for (int i = 0; i < GLTF.textures.Count; ++i)
-            {
-
-                TextureItem item = null;
-#if UNITY_EDITOR
-                var image = GLTF.GetImageFromTextureIndex(i);
-                if (imageBaseDir.IsUnderAssetsFolder
-                    && !string.IsNullOrEmpty(image.uri)
-                    && !image.uri.FastStartsWith("data:")
-                    )
-                {
-                    ///
-                    /// required SaveTexturesAsPng or SetTextureBaseDir
-                    ///
-                    var assetPath = imageBaseDir.Child(image.uri);
-                    var textureName = !string.IsNullOrEmpty(image.name) ? image.name : Path.GetFileNameWithoutExtension(image.uri);
-                    item = new TextureItem(i, assetPath, textureName);
-                }
-                else
-#endif
-                {
-                    item = new TextureItem(i, CreateTextureLoader(i));
-                }
-
-                AddTexture(item);
-            }
-        }
-
         /// <summary>
         /// Build unity objects from parsed gltf
         /// </summary>
@@ -578,23 +506,11 @@ namespace UniGLTF
             Schedulable.Create()
                 .AddTask(Scheduler.ThreadPool, () =>
                 {
-                    if (m_textures.Count == 0)
-                    {
-                        //
-                        // runtime
-                        //
-                        CreateTextureItems();
-                    }
-                    else
-                    {
-                        //
-                        // already CreateTextures(by assetPostProcessor or editor menu)
-                        //
-                    }
+                    m_materialFactory.Prepare(GLTF);
                 })
-                .ContinueWithCoroutine(Scheduler.ThreadPool, TexturesProcessOnAnyThread)
-                .ContinueWithCoroutine(Scheduler.MainThread, TexturesProcessOnMainThread)
-                .ContinueWithCoroutine(Scheduler.MainThread, LoadMaterials)
+                .ContinueWithCoroutine(Scheduler.ThreadPool, () => m_materialFactory.TexturesProcessOnAnyThread(GLTF, Storage))
+                .ContinueWithCoroutine(Scheduler.MainThread, () => m_materialFactory.TexturesProcessOnMainThread(GLTF))
+                .ContinueWithCoroutine(Scheduler.MainThread, () => m_materialFactory.LoadMaterials(GLTF))
                 .OnExecute(Scheduler.ThreadPool, parent =>
                 {
                     // UniGLTF does not support draco
@@ -649,47 +565,6 @@ namespace UniGLTF
             yield break;
         }
 
-        IEnumerator TexturesProcessOnAnyThread()
-        {
-            using (MeasureTime("TexturesProcessOnAnyThread"))
-            {
-                foreach (var x in GetTextures())
-                {
-                    x.ProcessOnAnyThread(GLTF, Storage);
-                    yield return null;
-                }
-            }
-        }
-
-        IEnumerator TexturesProcessOnMainThread()
-        {
-            using (MeasureTime("TexturesProcessOnMainThread"))
-            {
-                foreach (var x in GetTextures())
-                {
-                    yield return x.ProcessOnMainThreadCoroutine(GLTF);
-                }
-            }
-        }
-
-        IEnumerator LoadMaterials()
-        {
-            using (MeasureTime("LoadMaterials"))
-            {
-                if (GLTF.materials == null || !GLTF.materials.Any())
-                {
-                    AddMaterial(MaterialImporter.CreateMaterial(0, null, false, GetTexture));
-                }
-                else
-                {
-                    for (int i = 0; i < GLTF.materials.Count; ++i)
-                    {
-                        AddMaterial(MaterialImporter.CreateMaterial(i, GLTF.materials[i], GLTF.MaterialHasVertexColor(i), GetTexture));
-                    }
-                }
-            }
-            yield return null;
-        }
 
         IEnumerator BuildMesh(MeshImporter.MeshContext x, int i)
         {
@@ -793,45 +668,6 @@ namespace UniGLTF
         public GameObject Root;
         public List<Transform> Nodes = new List<Transform>();
 
-        List<TextureItem> m_textures = new List<TextureItem>();
-        public IList<TextureItem> GetTextures()
-        {
-            return m_textures;
-        }
-        public TextureItem GetTexture(int i)
-        {
-            if (i < 0 || i >= m_textures.Count)
-            {
-                return null;
-            }
-            return m_textures[i];
-        }
-        public void AddTexture(TextureItem item)
-        {
-            m_textures.Add(item);
-        }
-
-        List<Material> m_materials = new List<Material>();
-        public void AddMaterial(Material material)
-        {
-            var originalName = material.name;
-            int j = 2;
-            while (m_materials.Any(x => x.name == material.name))
-            {
-                material.name = string.Format("{0}({1})", originalName, j++);
-            }
-            m_materials.Add(material);
-        }
-        public IList<Material> GetMaterials()
-        {
-            return m_materials;
-        }
-        public Material GetMaterial(int index)
-        {
-            if (index < 0) return null;
-            if (index >= m_materials.Count) return null;
-            return m_materials[index];
-        }
 
         public List<MeshWithMaterials> Meshes = new List<MeshWithMaterials>();
         public void ShowMeshes()
@@ -866,7 +702,7 @@ namespace UniGLTF
         protected virtual IEnumerable<UnityEngine.Object> ObjectsForSubAsset()
         {
             HashSet<Texture2D> textures = new HashSet<Texture2D>();
-            foreach (var x in m_textures.SelectMany(y => y.GetTexturesForSaveAssets()))
+            foreach (var x in MaterialFacotry.GetTextures().SelectMany(y => y.GetTexturesForSaveAssets()))
             {
                 if (!textures.Contains(x))
                 {
@@ -874,7 +710,7 @@ namespace UniGLTF
                 }
             }
             foreach (var x in textures) { yield return x; }
-            foreach (var x in m_materials) { yield return x; }
+            foreach (var x in MaterialFacotry.GetMaterials()) { yield return x; }
             foreach (var x in Meshes) { yield return x.Mesh; }
             foreach (var x in AnimationClips) { yield return x; }
         }
@@ -1065,7 +901,7 @@ namespace UniGLTF
                 AssetDatabase.Refresh();
             }
 
-            CreateTextureItems(prefabParentDir);
+            m_materialFactory.CreateTextureItems(GLTF, prefabParentDir);
         }
         #endregion
 #endif
