+++ Assets/UniGLTF/Runtime/UniGLTF/MaterialFacotry.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bafc2eb5323419940a4991a8045bedc3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
