+++ Assets/UniGLTF/Runtime/UniGLTF/MaterialFacotry/MaterialFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49debef8f9942e8478c291dc7a41870c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
